Move Transition repository methods to Complaint repository
