+++ src/AppServices/Complaints/Dto/ComplaintViewDto.cs
@@ -1,6 +1,5 @@
 ﻿using Cts.AppServices.Attachments;
 using Cts.AppServices.ComplaintActions;
-using Cts.AppServices.ComplaintTransitions;
 using Cts.AppServices.Offices;
 using Cts.AppServices.Staff;
 using Cts.Domain.Entities.Complaints;
