--- src/Domain/Entities/ComplaintTransitions/IComplaintTransitionRepository.cs
@@ -1,3 +1,0 @@
-﻿namespace Cts.Domain.Entities.ComplaintTransitions;
-
-public interface IComplaintTransitionRepository : IRepository<ComplaintTransition, Guid> { }
