--- src/LocalRepository/Repositories/LocalComplaintTransitionRepository.cs
@@ -1,12 +1,0 @@
-﻿using Cts.Domain.Entities.ComplaintTransitions;
-using Cts.TestData;
-
-namespace Cts.LocalRepository.Repositories;
-
-/// <inheritdoc cref="IComplaintTransitionRepository" />
-public sealed class LocalComplaintTransitionRepository :
-    BaseRepository<ComplaintTransition, Guid>, 
-    IComplaintTransitionRepository
-{
-    public LocalComplaintTransitionRepository() : base(ComplaintTransitionData.GetComplaintTransitions) { }
-}
