+++ src/WebApp/Platform/Services/DataStores.cs
@@ -24,7 +24,6 @@ public static class DataStores
             // Uses local static data if no database is built.
             services.AddSingleton<IActionTypeRepository, LocalActionTypeRepository>();
             services.AddSingleton<IComplaintRepository, LocalComplaintRepository>();
-            services.AddSingleton<IComplaintTransitionRepository, LocalComplaintTransitionRepository>();
             services.AddSingleton<IConcernRepository, LocalConcernRepository>();
             services.AddSingleton<IOfficeRepository, LocalOfficeRepository>();
             services.AddSingleton<IUserRepository, LocalUserRepository>();
@@ -45,7 +44,6 @@ public static class DataStores
 
             services.AddScoped<IActionTypeRepository, ActionTypeRepository>();
             services.AddScoped<IComplaintRepository, ComplaintRepository>();
-            services.AddScoped<IComplaintTransitionRepository, ComplaintTransitionRepository>();
             services.AddScoped<IConcernRepository, ConcernRepository>();
             services.AddScoped<IOfficeRepository, OfficeRepository>();
             services.AddScoped<IUserRepository, UserRepository>();
